Fix header file conflicts and type mismatches
- Add ifndef guards for duplicate struct definitions (HiScore, UVect, TireState)
- Add ifndef guards for conflicting macros (MAX_GAME_CNT, MAX_HI_CNT, NSCORES, etc.)
- Fix type mismatches between header declarations and source definitions
  - end_game_flag/lap_flag: s8 -> s32
  - GetElapsedTime: u32 -> s32
  - ShowHiScore track parameter: s16 -> s32
- Remove conflicting arcade alias macros from camera.h
- Add catchup field to CarPhysics struct
- Fix we_control_flag -> we_control in model.c
- Add forward declarations for vec_* functions in physics.c
- Rename ONE_SEC to ONE_SEC_MS in hiscore.h to avoid conflict

All 120 C files now compile without errors.

+++ src/game/checkpoint.c
@@ -104,8 +104,8 @@ s8  cancel_mpath_lap;
 f32 lap_loop_distance;
 s16 close_chkpnt;
 s16 path_point_index;
-s8  end_game_flag;
-s8  lap_flag;
+s32 end_game_flag;
+s32 lap_flag;
 
 /* Track/checkpoint data arrays */
 CheckPoint TrackCPs[MAX_TRACKS][MAX_CHECKPOINTS];
